pyproject.toml
[tool.poetry]
name = "discordbottemplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
discord = "^1.0.1"
replit = "^2.0.0"
python-dotenv = "^0.15.0"
termcolor = "^1.1.0"
youtube_dl = "^2021.12.17"
"discord.py" = "^1.7.3"
discord-voice = "^0.0.0"
PyNaCl = "^1.5.0"
tasks = "^2.8.0"
metronome-loop = "^0.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
